Refactor: Move 2FA modal and form binding functions to login.ts for better organization
